Fix an issue due to unexpected gcc behavior
gcc will do sign extension for below case:

"
struct {
	uint32_t lrca:20;
}a;

unsigned long val = a.lrca << 12;
"

if MSB of "lrca" is 1, the result of "val" is like 0xffffffffxxxxxxxx;

Doing explicit type converting fixes this.

This can fix #916, that with 1G aperture, starting the third VM will have
problems.

+++ drivers/gpu/drm/i915/vgt/execlists.c
@@ -819,7 +819,7 @@ static int vgt_create_shadow_pages(struct vgt_device *vgt, struct execlist_conte
 		return -1;
 	}
 
-	g_gma = el_ctx->guest_context.lrca << GTT_PAGE_SHIFT;
+	g_gma = ((unsigned long)el_ctx->guest_context.lrca) << GTT_PAGE_SHIFT;
 	s_gma = aperture_2_gm(vgt->pdev, hpa);
 
 	el_ctx->shadow_lrca = s_gma >> GTT_PAGE_SHIFT;
